+++ R/utils.R
@@ -1,38 +1,3 @@
-#' @title Segmentize an observation-by-period data frame
-#' @description Convert an observation-by-period data frame into a matrix of line segment coordinates, where each row represents a line segment connecting two time points from the original data.
-#' @param data A data frame containing observation-by-period data, where the only columns represent sequentially ordered time period values for each observation. Some values can be missing.
-#' @return A five-column matrix containing: the row from the original data frame, and x1, x2, y1, y2 positions for each segment.
-#' @examples
-#' data(gdp)
-#' head(segmentize(gdp))
-#' @importFrom stats embed na.omit
-#' @export
-segmentize <- function(data) {
-    # create structure that is row, x1, x2, y1, y2
-    # for every pair of contiguous points
-    
-    if (ncol(data) == 2) {
-        pairsmat <- matrix(1:2, nrow = 1L)
-    } else {
-        pairsmat <- embed(seq_len(ncol(data)), 2)[,2:1]
-    }
-    # output
-    out <- matrix(NA_real_, nrow = nrow(data) * nrow(pairsmat), ncol = 5L)
-    k <- 1L
-    for (i in seq_len(nrow(data))) {
-        for (j in seq_len(nrow(pairsmat))) {
-            out[k,] <- c(i, 
-                         pairsmat[j,1], 
-                         pairsmat[j,2], 
-                         data[i, pairsmat[j,1]], 
-                         data[i, pairsmat[j,2]])
-            k <- k + 1L
-        }
-    }
-    # return, dropping missing values
-    na.omit(out)
-}
-
 # function for finding consecutive indices
 # from: http://stackoverflow.com/a/16118320/2338862
 seqle <- function(x, incr=1) {