+++ NEWS.md
@@ -1,3 +1,8 @@
+# slopegraph 0.1.10
+
+* `segmentize()` gains new argument `na.span`, a logical indicating whether line segments should span periods with missing values. The default remains to not do this and simply omit segments when missing values are present. This argument is cascaded through to `slopegraph()` and `ggslopegraph()`. (h/t Joran Elias)
+* Drawing of left- and right-side observation labels has been made optional (by setting `labpos.left` or `labpos.right` to `NULL`).
+
 # slopegraph 0.1.9
 
 * Added a new function, `ggslopegraph()` that produces a ggplot2-based slopegraph.